+++ sdk/identity/identity/src/msal/nodeFlows/msalClientSecret.ts
@@ -33,7 +33,8 @@ export class MsalClientSecret extends MsalNode {
       const result = await this.confidentialApp!.acquireTokenByClientCredential({
         scopes,
         correlationId: options.correlationId,
-        azureRegion: this.azureRegion
+        azureRegion: this.azureRegion,
+        authority: options.authority
       });
       // The Client Credential flow does not return an account,
       // so each time getToken gets called, we will have to acquire a new token through the service.